feat: set/change some column data types and constrains

+++ packages/llstarscreamll/Company/database/migrations/2014_10_11_000001_create_sub_cost_centers_table.php
@@ -20,7 +20,7 @@ class CreateSubCostCentersTable extends Migration
     {
         Schema::create('sub_cost_centers', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedInteger('cost_center_id');
+            $table->unsignedBigInteger('cost_center_id');
             $table->string('code')->unique();
             $table->string('name');
             $table->timestamps();
